Days 9 and 10. Undid the Vector int to long change (it broked some stuff).

+++ AoCLib/Vector2.cs
@@ -10,10 +10,10 @@ namespace Draco18s.AoCLib {
 		public static readonly Vector2 UP = new Vector2(0, -1);
 		public static readonly Vector2 DOWN = new Vector2(0, 1);
 
-		public readonly long x;
-		public readonly long y;
+		public readonly int x;
+		public readonly int y;
 		public double magnitude => Math.Sqrt(x * x + y * y);
-		public Vector2(long _x, long _y) {
+		public Vector2(int _x, int _y) {
 			x = _x;
 			y = _y;
 		}
@@ -26,14 +26,14 @@ namespace Draco18s.AoCLib {
 		public static Vector2 Parse(string val, char split)
 		{
 			string[] vals = val.Split(split);
-			return new Vector2(long.Parse(vals[0]), long.Parse(vals[1]));
+			return new Vector2(int.Parse(vals[0]), int.Parse(vals[1]));
 		}
 
-		public static Vector2 operator *(Vector2 a, long b)
+		public static Vector2 operator *(Vector2 a, int b)
 		{
 			return new Vector2(a.x * b, a.y * b);
 		}
-		public static Vector2 operator /(Vector2 a, long b)
+		public static Vector2 operator /(Vector2 a, int b)
 		{
 			return new Vector2(a.x / b, a.y / b);
 		}
